requirements.txt
flask==3.1.1
python-dateutil==2.9.0
peewee==3.18.1
requests~=2.32.4